Now compiles on Solaris with GSL and GCC 3.4.3.
- Added implementation of isinf() in utility.h, since isinf
  is not part of standard C++. The new function is isInfinity().
- Change of a variable's type in GenesisParserWrapper from unsigned
  to signed. This gets rid of a warning.


git-svn-id: https://svn.code.sf.net/p/moose/code/moose/trunk@604 d1858a74-f715-0410-99c9-c66fe9cce3d6

Former-commit-id: ed683acfc34998d85c026d37598897f59bf9e43c

+++ genesis_parser/GenesisParserWrapper.cpp
@@ -3457,8 +3457,8 @@ int do_findchar(int argc, const char** const argv, Id s ){
 	}
 	string str1 = argv[1];
 	string str2 = argv[2];
-	size_t pos = str1.find(str2, 0); 	
-	if (pos == string::npos)
+	int pos = str1.find(str2, 0); 	
+	if ( pos == ( int )( string::npos ) )
 		pos = -1;
 	return pos;
 }
